Status Code 204

+++ back/src/ProEventos.API/Controllers/EventosController.cs
@@ -25,7 +25,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventosAsync(true);
-                if(eventos == null)return NotFound("Nenhum evento encontrado.");
+                if(eventos == null)return NoContent();
 
 
                 return Ok(eventos);
@@ -44,7 +44,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetEventoByIdAsync(id, true);
-                if(evento == null)return NotFound("Evento por Id não encontrado.");
+                if(evento == null)return NoContent();
                 return Ok(evento);
             }
             catch (Exception ex)
@@ -61,7 +61,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-                if(evento == null)return NotFound("Eventos por tema não encontrados.");
+                if(evento == null)return NoContent();
                 return Ok(evento);
             }
             catch (Exception ex)
@@ -111,9 +111,12 @@ namespace ProEventos.API.Controllers
         {
             try
             {
+                var evento  = await _eventoService.GetEventoByIdAsync(id, true);
+                if(evento==null) return NoContent();
+
                 return await _eventoService.DeleteEvento(id) ?
                         Ok("Deletado") : 
-                        BadRequest("Evento não deletado");
+                        throw new Exception("Ocorreu um problema não especifico ao tentar deletar Evento.");
             }
             catch (Exception ex)
             {
